`Added "btn-json" button and related functionality in chatbot`
` `

`A new button "btn-json" has been added to the chatbot footer in Chatbot.cshtml, styled with "btn" and "btn-outline-success" classes. Clicking this button sets the message input field to "Give me JSON" and triggers a send event. In chat.js, the button is disabled during message sending and re-enabled after the bot has processed the message or if an error occurs during processing.`